+++ src/service/api/article/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const {validateArticle} = require(`./validate-article`);
+const {ArticlesService} = require(`./articleService`);
+const {articleRouter} = require(`./article-router`);
+
+module.exports = {
+  validateArticle,
+  ArticlesService,
+  articleRouter,
+};
