+++ src/service/api/category/category-router.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const {Router} = require('express');
-const {HTTP_CODE} = require('../../constants');
+const {HTTP_CODE} = require('../../../constants');
 
 function createCategoryRouter(app, categoryService) {
   const categoriesRouter = new Router();
