+++ src/service/api/category/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const {categoryRouter} = require(`./category-router`);
+const {CategoryService} = require(`./CategoryService`);
+
+module.exports = {
+  categoryRouter,
+  CategoryService,
+};
