--- src/service/data-service/index.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  CategoryService: require('./CategoryService'),
-  ArticleService: require('./ArticleService'),
-}
