--- src2/controllers/ArticleController.js
@@ -1,1 +1,0 @@
-// Принимает HTTP запросы
